Rotate works

+++ Assets/_scripts/RotateOnClick.cs
@@ -9,12 +9,12 @@ public class RotateOnClick : MonoBehaviour
     #region Serialized Private Fields
     [SerializeField] private GameObject rubik;
     [SerializeField] private RubiksCube.Section sector;
-    [SerializeField] private int specifier;
+    [SerializeField] private float specifier;
     [SerializeField] private bool negative;
     #endregion
 
     #region Public Method
-    public void SetValues(RubiksCube.Section sec, int spec, bool neg)
+    public void SetValues(RubiksCube.Section sec, float spec, bool neg)
     {
         sector = sec;
         specifier = spec;
